+++ src/main/java/ajou/mse/dimensionguard/repository/redis/RedisSkillInfoRepository.java
@@ -3,5 +3,5 @@ package ajou.mse.dimensionguard.repository.redis;
 import ajou.mse.dimensionguard.dto.redis.RedisSkillInfo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RedisSkillInfoRepository extends CrudRepository<RedisSkillInfo, Integer> {
+public interface RedisSkillInfoRepository extends CrudRepository<RedisSkillInfo, Long> {
 }
